backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pillow
pydantic
openai
requests